// src/generated/sovd.d.ts
export type webhooks = Record<string, never>;

export type components = Record<string, never>;

export type $defs = Record<string, never>;

export interface external {
// ...
  "commons/types.yaml": {
    paths: Record<string, never>;
    webhooks: Record<string, never>;
    components: {
      schemas: {
        AnyValue: string | number | boolean | unknown[] | Record<string, never>;
        /** @description The schema definition of the response. Condition: Only provided if the query parameter include-schema is true. */
        OpenApiSchema: external["https://raw.githubusercontent.com/OAI/OpenAPI-Specification/main/schemas/v3.1/schema.yaml"]["$defs"]["schema"];
        /** @description Generic definition of an error */
        DataError: {
          /**
           * Format: json-pointer
           * @description JSON Pointer pointing to the erroneous value
           */
          path: string;
          error: external["commons/errors.yaml"]["components"]["schemas"]["GenericError"];
        };
        /** @description EntityReference type */
        EntityReference: {
          /** @description Identifier of the entity */
          id: string;
          /** @description Name of the entity. */
          name: string;
          /** @description Identifier for translating the name */
          translation_id?: string;
          /**
           * Format: uri-reference
           * @description URI of the subordinate entity including {base_uri}.
           */
          href: string;
        };
        /**
         * @description The categories identData, currentData and sysInfo are predefined by SOVD. OEM-specific categories can be included alpha-numeric characters, hyphens, and underscores and have to start with x-<oem>
         * @example currentData
         */
        DataCategory: string;
        Value: {
          /** @description Unique identifier for a value */
          id: string;
          /** @description Retrieved value of the entity */
          data: Record<string, never>;
          metadata?: external["commons/types.yaml"]["components"]["schemas"]["ValueMetadata"];
          error?: external["commons/types.yaml"]["components"]["schemas"]["DataError"];
        };
        ListOfValues: {
          items?: external["commons/types.yaml"]["components"]["schemas"]["Value"][];
        };
        /** @description Description of a value */
        ValueMetadata: {
          /**
           * @description Unique identifier for a value
           * @example DriverWindow
           */
          id: string;
          /**
           * @description Name of the data value
           * @example Position of driver window
           */
          name: string;
          /** @description Identifier for translating the data’s name */
          translation_id?: string;
          category: external["commons/types.yaml"]["components"]["schemas"]["DataCategory"];
          /** @description The identifiers of the groups to which the resource belongs to. */
          groups?: string[];
        };
        /** @description Single data value */
        ReadValue: {
          /** @description Unique identifier for a value */
          id: string;
          /** @description The value of the data. The type of the value is defined by the schema. */
          data: external["commons/types.yaml"]["components"]["schemas"]["AnyValue"];
          /** @description Condition: Only set, if the value data represents an error and this attribute describes the error more precisely. */
          errors?: external["commons/errors.yaml"]["components"]["schemas"]["DataError"][];
          schema?: external["commons/types.yaml"]["components"]["schemas"]["OpenApiSchema"];
        };
      };
      responses: never;
      parameters: never;
      requestBodies: never;
      headers: never;
      pathItems: never;
    };
    $defs: Record<string, never>;
  };
// ...
  "logs/types.yaml": {
    paths: Record<string, never>;
    webhooks: Record<string, never>;
    components: {
      schemas: {
        /**
         * @example {
         *   "timestamp": "2021-07-20T00:00:04.387819Z",
         *   "context": {
         *     "type": "RFC5424",
         *     "host": "Linux",
         *     "process": "systemd",
         *     "pid": 1
         *   },
         *   "severity": "info",
         *   "msg": "Closed D-Bus User Message Bus Socket",
         *   "href": "https://sovd.server/v1/components/DrivingComputer/bulk-data/logs/server.log"
         * }
         */
        LogEntry: {
          /**
           * Format: date-time
           * @description Format: RFC3339 (UTC)
           */
          timestamp: string;
          /** @description Definition of the context of the log message identifying the source of the entry. */
          context: external["logs/types.yaml"]["components"]["schemas"]["Context"];
          severity: external["logs/types.yaml"]["components"]["schemas"]["Severity"];
          /** @description The actual log message (text) */
          msg: string;
          /**
           * Format: uri-reference
           * @description Reference to a URI for retrieving the underlying original log file in an application specific format
           */
          href?: string;
        };
        Context: external["logs/types.yaml"]["components"]["schemas"]["RFC5424Context"] | external["logs/types.yaml"]["components"]["schemas"]["AUTOSAR_DLTContext"];
        /**
         * @example {
         *   "type": "RFC5424",
         *   "host": "Linux",
         *   "process": "systemd",
         *   "pid": 1
         * }
         */
        RFC5424Context: {
          /** @constant */
          type: "RFC5424";
          host?: string;
          process?: string;
          pid?: number;
        };
        /**
         * @example {
         *   "type": "AUTOSAR_DLT",
         *   "session": "SES01",
         *   "session_id": "0x0001",
         *   "application_id": "WindowControl",
         *   "context_id": "STD0",
         *   "message_id": "msg000012"
         * }
         */
        AUTOSAR_DLTContext: {
          /** @constant */
          type: "AUTOSAR_DLT";
          session?: string;
          session_id?: string;
          application_id?: string;
          context_id?: string;
          message_id?: string;
        };
        /**
         * @description SOVD specific datatype defining log-levels
         * @enum {string}
         */
        Severity: "fatal" | "error" | "warn" | "info" | "debug";
        LogConfiguration: {
          /** @description Definition of the context of the log message. Parts not specified in the context are regarded as not set. */
          context: external["logs/types.yaml"]["components"]["schemas"]["Context"];
          /** @description The severity specifies the log-level for the given context. */
          severity: external["logs/types.yaml"]["components"]["schemas"]["Severity"];
        };
      };
      responses: never;
      parameters: never;
      requestBodies: never;
      headers: never;
      pathItems: never;
    };
    $defs: Record<string, never>;
  };
// ...
}